refactor!: separate type exports from useexploremode hook to avoid test import issues (#414)
