+++ src/config/entities.ts
@@ -19,7 +19,7 @@ import {
 import { HeroTitle } from "../components/common/Title/title";
 import { FooterProps } from "../components/Layout/components/Footer/footer";
 import { HeaderProps } from "../components/Layout/components/Header/header";
-import { ExploreMode } from "../hooks/useExploreMode";
+import { ExploreMode } from "../hooks/useExploreMode/types";
 import { AuthState } from "../providers/authentication/auth/types";
 import { UserProfile } from "../providers/authentication/authentication/types";
 import { ProviderId } from "../providers/authentication/common/types";
