+++ src/hooks/useExploreMode/types.ts
@@ -1,0 +1,7 @@
+export enum EXPLORE_MODE {
+  CS_FETCH_CS_FILTERING = "CS_FETCH_CS_FILTERING",
+  SS_FETCH_CS_FILTERING = "SS_FETCH_CS_FILTERING",
+  SS_FETCH_SS_FILTERING = "SS_FETCH_SS_FILTERING",
+}
+
+export type ExploreMode = EXPLORE_MODE;
